STEP 4.32 DONE

+++ script/HelperConfig.s.sol
@@ -20,7 +20,7 @@ contract HelperConfig is Script, CodeConstants {
         uint256 interval;
         address vrfCoordinator;
         bytes32 gasLane;
-        uint64 subscriptionId;
+        uint256 subscriptionId;
         uint32 callbackGasLimit;
     }
